requirements.txt
streamlit==1.39.0
tensorflow-cpu==2.13.0
numpy==1.24.4
pillow==10.4.0
protobuf==4.25.3

